Some dockerfile SBOMS fail validation
Problem:
Validator complains about 'author' appearing before 'group' or name'.

Solution:
Enforce latest versions of syft, cdx cmds and check ordering guarantees
with Elementtree python module.
Upgrade to cyclonedx 1.4.
Create Dockerfile to enforce dependency versions and enable local
testing.
Remove xmllinting of cyclonedx as upstream now handles this.
Executing xmllint was failing to invalidate the xml file because of the
need to execute in a pipeline.
sbom_scraper.sh can now be executed from any location and is not tied to
being executed in the repo.

Signed-off-by: Paul Hewlett <phewlett76@gmail.com>